Move 'extract' actions to RefactorProcessor
